southern-code-django-challenge-main/requirements.txt
mysql-connector-python==8.0.33
dj-database-url==2.0.0
dj-static==0.0.6
django==3.2
django-ajax-datatable==4.4.5
django-automatic-crud==1.2.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-filter==23.2
django-import-export==3.2.0
djangorestframework==3.14.0
drf-yasg==1.21.6
Faker==18.13.0
gunicorn==20.1.0
requests==2.30.0
whitenoise==6.4.0
requests==2.30.0
djangorestframework==3.14.0


